Move import logic out to partial class
